refactor(server): storage template core
