+++ server/src/domain/storage-template/index.ts
@@ -1,2 +1,1 @@
-export * from './storage-template.core';
 export * from './storage-template.service';
